requirements.txt
pbr
xopen
biopython<=1.76
tabulate
plotille
